move utils to useWysiwyg
