+++ platforms/web/src/App.tsx
@@ -31,7 +31,6 @@ import quoteImage from './images/quote.svg';
 import indentImage from './images/indent.svg';
 import unindentImage from './images/unindent.svg';
 import { Wysiwyg, WysiwygEvent } from '../lib/types';
-import { SuggestionPattern } from '../generated/wysiwyg';
 
 type ButtonProps = {
     onClick: MouseEventHandler<HTMLButtonElement>;
@@ -55,33 +54,6 @@ function Button({ onClick, imagePath, alt, state }: ButtonProps) {
             <img alt={alt} src={imagePath} />
         </button>
     );
-}
-
-function getSuggestionChar(suggestion: SuggestionPattern) {
-    const suggestionMap = ['@', '#', '/'];
-    return suggestionMap[suggestion.key];
-}
-
-function getSuggestionType(suggestion: SuggestionPattern) {
-    switch (suggestion.key) {
-        case 0:
-        case 1:
-            return 'mention';
-        case 2:
-            return 'command';
-        default:
-            return 'unknown';
-    }
-}
-
-// TODO use this when passing the output out from the hook => this way we can
-// keep the state consistent, but use 'MappedSuggestion' type on the React side
-function mapSuggestion(suggestion: SuggestionPattern) {
-    return {
-        ...suggestion,
-        keyChar: getSuggestionChar(suggestion),
-        suggestionType: getSuggestionType(suggestion),
-    };
 }
 
 function App() {
@@ -215,7 +187,7 @@ function App() {
                         <button type="button" onClick={(_e) => wysiwyg.clear()}>
                             clear
                         </button>
-                        {isMention(suggestion) && (
+                        {suggestion && suggestion.type === 'mention' && (
                             <button
                                 type="button"
                                 onClick={(_e) =>
@@ -225,10 +197,10 @@ function App() {
                                     )
                                 }
                             >
-                                Add {getSuggestionChar(suggestion)}mention
+                                Add {suggestion.keyChar}mention
                             </button>
                         )}
-                        {isCommand(suggestion) && (
+                        {suggestion && suggestion.type === 'command' && (
                             <button
                                 type="button"
                                 onClick={(_e) => wysiwyg.command('/spoiler ')}
